Plotting updates
Add tie offset if there are overlapping lines - so can see all data

Changed how get latest_end_date - it is not the max of the end dates it is the latest.  This is used for ordering of the legend entries.

Adjusted how set axis limits - to add a month only if the total range is over 4 months.  Otherwise just add a day to each end of the range for display aesthetics.