Minor Updates
Added hierarchical_task_name to Task, so we can check them when we add TaskLists to a TaskListArray - to double check that the numbers are always referring to the same task.  (In case we changed the order of the task somehow).

Removed tabs from csv output of TaskListArray - the tabs were being parsed and displayed in the plots...

Fixed the spelling of "hierarchical".